requirements.txt
google-cloud-vision==0.41.0
google-cloud-storage==1.19.1